LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.fillStyle.parse.hsl-clamp-2.worker-expected.txt


FAIL 2d assert_equals: Red channel of the pixel at (50, 25) expected 127 but got 255
